+++ __docs__/wonder-blocks-dropdown/option-item.stories.tsx
@@ -55,7 +55,9 @@ export default {
     decorators: [
         (Story): React.ReactElement<React.ComponentProps<typeof View>> => (
             <View style={styles.example}>
-                <Story />
+                <div role="listbox" aria-label="Example">
+                    <Story />
+                </div>
             </View>
         ),
     ],