+++ packages/wonder-blocks-labeled-field/src/components/labeled-field.tsx
@@ -3,8 +3,7 @@ import {StyleSheet} from "aphrodite";
 import WarningCircle from "@phosphor-icons/core/bold/warning-circle-bold.svg";
 
 import {View, addStyle, StyleType} from "@khanacademy/wonder-blocks-core";
-import {Strut} from "@khanacademy/wonder-blocks-layout";
-import {semanticColor, spacing} from "@khanacademy/wonder-blocks-tokens";
+import {semanticColor, sizing} from "@khanacademy/wonder-blocks-tokens";
 import {LabelMedium, LabelSmall} from "@khanacademy/wonder-blocks-typography";
 import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
 
@@ -55,9 +54,15 @@ type Props = {
      */
     errorMessage?: React.ReactNode;
     /**
-     * Custom styles for the labeled field container.
+     * Custom styles for the elements of LabeledField. Useful if there are
+     * specific cases where spacing between elements needs to be customized.
      */
-    style?: StyleType;
+    styles?: {
+        root?: StyleType;
+        label?: StyleType;
+        description?: StyleType;
+        error?: StyleType;
+    };
     /**
      * A unique id to use as the base of the ids for the elements within the component.
      * Here is how the id is used for the different elements in the component:
@@ -109,7 +114,7 @@ const StyledSpan = addStyle("span");
 export default function LabeledField(props: Props) {
     const {
         field,
-        style,
+        styles: stylesProp,
         label,
         id,
         required,
@@ -143,7 +148,14 @@ export default function LabeledField(props: Props) {
         return (
             <React.Fragment>
                 <LabelMedium
-                    style={[styles.textWordBreak, styles.label]}
+                    style={[
+                        styles.textWordBreak,
+                        styles.label,
+                        description
+                            ? styles.labelWithDescription
+                            : styles.labelWithNoDescription,
+                        stylesProp?.label,
+                    ]}
                     tag="label"
                     htmlFor={fieldId}
                     testId={testId && `${testId}-label`}
@@ -152,7 +164,6 @@ export default function LabeledField(props: Props) {
                     {label}
                     {isRequired && requiredIcon}
                 </LabelMedium>
-                <Strut size={spacing.xxxSmall_4} />
             </React.Fragment>
         );
     }
@@ -165,13 +176,16 @@ export default function LabeledField(props: Props) {
         return (
             <React.Fragment>
                 <LabelSmall
-                    style={[styles.textWordBreak, styles.description]}
+                    style={[
+                        styles.textWordBreak,
+                        styles.description,
+                        stylesProp?.description,
+                    ]}
                     testId={testId && `${testId}-description`}
                     id={descriptionId}
                 >
                     {description}
                 </LabelSmall>
-                <Strut size={spacing.xxxSmall_4} />
             </React.Fragment>
         );
     }
@@ -185,6 +199,7 @@ export default function LabeledField(props: Props) {
                         errorMessage
                             ? styles.errorSectionWithContent
                             : undefined,
+                        stylesProp?.error,
                     ]}
                     id={errorId}
                     testId={testId && `${testId}-error`}
@@ -239,10 +254,9 @@ export default function LabeledField(props: Props) {
     }
 
     return (
-        <View style={style}>
+        <View style={stylesProp?.root}>
             {renderLabel()}
             {maybeRenderDescription()}
-            <Strut size={spacing.xSmall_8} />
             {renderField()}
             {maybeRenderError()}
         </View>
@@ -253,15 +267,22 @@ const styles = StyleSheet.create({
     label: {
         color: semanticColor.text.primary,
     },
+    labelWithDescription: {
+        paddingBlockEnd: sizing.size_050,
+    },
+    labelWithNoDescription: {
+        paddingBlockEnd: sizing.size_150,
+    },
     description: {
         color: semanticColor.text.secondary,
+        paddingBlockEnd: sizing.size_150,
     },
     errorSection: {
         flexDirection: "row",
-        gap: spacing.xSmall_8,
+        gap: sizing.size_100,
     },
     errorSectionWithContent: {
-        paddingTop: spacing.small_12,
+        paddingBlockStart: sizing.size_150,
     },
     error: {
         color: semanticColor.status.critical.foreground,