+++ packages/wonder-blocks-tokens/src/tokens/semantic-color.ts
@@ -161,7 +161,7 @@ export const semanticColor = {
      * Default icon colors that change in context (like actions).
      */
     icon: {
-        primary: color.fadedOffBlack64,
+        primary: color.fadedOffBlack72,
         secondary: color.offBlack,
         inverse: color.white,
         action: color.blue,