+++ packages/wonder-blocks-icon-button/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-icon-button
 
+## 6.1.4
+
+### Patch Changes
+
+- 999101f: Fix IconButton theme to use semanticColor.icon tokens
+- 1d7be37: Use pnpm catalog to pin dependency versions across packages
+- Updated dependencies [1d7be37]
+- Updated dependencies [999101f]
+    - @khanacademy/wonder-blocks-clickable@6.1.4
+    - @khanacademy/wonder-blocks-theming@3.2.1
+    - @khanacademy/wonder-blocks-core@12.2.1
+    - @khanacademy/wonder-blocks-icon@5.1.3
+    - @khanacademy/wonder-blocks-tokens@5.1.1
+
 ## 6.1.3
 
 ### Patch Changes